requirements.txt
openai==1.30.4
python-dotenv==1.0.1
pypdf==4.2.0
langchain_community==0.2.1
llama_index==0.10.40
langchain_openai==0.1.8
langchain_chroma==0.1.1
streamlit
streamlit_extras
flask